Update HMM-Poisson model.
